+++ Assets/Scripts/Interface/IDamagable.cs
@@ -1,7 +1,6 @@
 ﻿using UnityEngine;
 
 public interface IDamagable
-{
-    
-    void TakeDamage(int damage);
+{   
+    void TakeDamage(float damage);
 }
